requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langgraph
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai
tavily-python